requirements.txt
-i https://pypi.org/simple
certifi==2022.9.24; python_version >= '3.6'
commonmark==0.9.1
dotmap==1.3.30
enum-compat==0.0.3
influxdb-client==1.32.0
meshtastic==1.3.36
pexpect==4.8.0
protobuf==4.21.6; python_version >= '3.7'
ptyprocess==0.7.0
pydantic[dotenv]==1.10.2
pygatt==4.0.5; platform_system == 'Linux'
pygments==2.13.0; python_version >= '3.6'
pypubsub==4.0.3; python_version >= '3.3' and python_version < '4'
pyqrcode==1.2.1
pyserial==3.5
pytap2==2.2.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0
pyyaml==6.0; python_version >= '3.6'
reactivex==4.0.4; python_version >= '3.7' and python_version < '4'
rich==12.5.1
setuptools==65.4.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tabulate==0.8.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
timeago==1.0.16
typing-extensions==4.3.0; python_version >= '3.7'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
